requirements.txt
streamlit==1.56.0

langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.40
langchain-groq==0.1.9
langchain-chroma==0.1.4
langchain-text-splitters==0.2.4

chromadb==0.5.23

sentence-transformers==5.4.0

PyMuPDF==1.27.2.2

python-dotenv==1.2.2

groq==0.37.1